# Optimize simplifying paths for faster project generation