@brentleyjones 2020-05-08
This makes our project generation 1-2 seconds faster (~9%).